Fixed ID10T error... and got kestrel working again
