+++ src/OmniSharp.Http/OmniSharpHttpEnvironment.cs
@@ -1,0 +1,7 @@
+namespace OmniSharp.Http
+{
+    public class OmniSharpHttpEnvironment
+    {
+        public int Port { get; set; }
+    }
+}
