+++ src/OmniSharp.Http/Startup.cs
@@ -20,23 +20,21 @@ namespace OmniSharp.Http
         private readonly IOmniSharpEnvironment _environment;
         private readonly IEventEmitter _eventEmitter;
         private readonly ISharedTextWriter _writer;
-        private readonly OmniSharpHttpEnvironment _httpEnvironment;
         private readonly IConfigurationRoot _configuration;
         // private OmniSharpWorkspace _workspace;
         private CompositionHost _compositionHost;
 
-        public Startup()
+        public Startup(IOmniSharpEnvironment environment, IEventEmitter eventEmitter, ISharedTextWriter writer)
         {
-            _environment = Program.Instance._environment;
-            _eventEmitter = NullEventEmitter.Instance;
-            _writer = Program.Instance._sharedTextWriter;
-            _httpEnvironment = new OmniSharpHttpEnvironment { Port = Program.Instance._serverPort };
-            _configuration = new OmniSharpConfigurationBuilder(_environment).Build();
+            _environment = environment;
+            _eventEmitter = eventEmitter;
+            _writer = writer;
+            _configuration = new OmniSharpConfigurationBuilder(environment).Build();
         }
 
         public IServiceProvider ConfigureServices(IServiceCollection services)
         {
-            var serviceProvider = OmniSharpMefBuilder.CreateDefaultServiceProvider(_configuration);
+            var serviceProvider = OmniSharpMefBuilder.CreateDefaultServiceProvider(_configuration, services);
             var mefBuilder = new OmniSharpMefBuilder(serviceProvider, _environment, _writer, _eventEmitter);
             var compositionHost = mefBuilder.Build();
             _compositionHost = compositionHost;
@@ -47,8 +45,12 @@ namespace OmniSharp.Http
             IApplicationBuilder app,
             IServiceProvider serviceProvider,
             ILoggerFactory loggerFactory,
+            IEventEmitter eventEmitter,
+            ISharedTextWriter writer,
+            OmniSharpHttpEnvironment httpEnvironment,
             IOptionsMonitor<OmniSharpOptions> options)
         {
+            var workspace = _compositionHost.GetExport<OmniSharpWorkspace>();
             var logger = loggerFactory.CreateLogger<Startup>();
             loggerFactory.AddConsole((category, level) =>
             {
@@ -64,13 +66,13 @@ namespace OmniSharp.Http
 
             app.UseRequestLogging();
             app.UseExceptionHandler("/error");
-            app.UseMiddleware<EndpointMiddleware>();
-            app.UseMiddleware<StatusMiddleware>();
+            app.UseMiddleware<EndpointMiddleware>(_compositionHost);
+            app.UseMiddleware<StatusMiddleware>(workspace);
             app.UseMiddleware<StopServerMiddleware>();
 
             new OmniSharpWorkspaceInitializer(serviceProvider, _compositionHost, _configuration, logger).Initialize();
 
-            logger.LogInformation($"Omnisharp server running on port '{_httpEnvironment.Port}' at location '{_environment.TargetDirectory}' on host {_environment.HostProcessId}.");
+            logger.LogInformation($"Omnisharp server running on port '{httpEnvironment.Port}' at location '{_environment.TargetDirectory}' on host {_environment.HostProcessId}.");
         }
     }
 }
